+++ resources/views/slides/show.blade.php
@@ -35,7 +35,12 @@
 							</tr>
 							@foreach ($collection->items as $item)
 							<tr>
-								<td><img src="{!! $item->getImagePath(); !!}" alt="náhľad" class="img-responsive" ></td>
+								<td>
+									<img src="{!! $item->getImagePath(); !!}" 
+									srcset="{!! $item->getImageSrcset(); !!}" 
+									alt="náhľad" class="img-responsive"
+									onerror="this.onerror=null;this.srcset=''"
+									/></td>
 								<td>
 									<a href="{!! URL::to('item/' . $item->id . '/edit' ) !!}">{!! $item->author !!} - {!! $item->title !!}</a>
 								</td>